Users can edit and delete their own posts only

+++ resources/views/home.blade.php
@@ -14,12 +14,14 @@
                 <div class="card-body">
                     {{ $rev->content }}
                 </div>
-                <div class="card-body">
-                    <form style="float:left;margin-right:10px;"action="{!!route('contents.edit', [$rev->id])!!}"><input type="submit" value="EDIT" /></form>
-                    {!! Form::open(['route'=>['contents.destroy',$rev->content_id], 'method' => 'delete'])!!}
-                    {!! Form::button('DELETE', ['type' => 'submit', 'onclick' => "return confirm ('Are you sure?')"])!!}
-                    {!! Form::close()!!}
-                </div>
+                @if(Auth::check() && Auth::user()->id == $rev->user_id)
+                      <div class="card-body">
+                          <form style="float:left;margin-right:10px;"action="{!!route('contents.edit', [$rev->id])!!}"><input type="submit" value="EDIT" /></form>
+                          {!! Form::open(['route'=>['contents.destroy',$rev->content_id], 'method' => 'delete'])!!}
+                          {!! Form::button('DELETE', ['type' => 'submit', 'onclick' => "return confirm ('Are you sure?')"])!!}
+                          {!! Form::close()!!}
+                      </div>
+                @endif
             </div><br/><br/>
             <?php $test="$rev->content_id";?>
           @endforeach
